Split StreamBackoff into a separate utils module
