+++ kube-runtime/src/utils/mod.rs
@@ -1,4 +1,7 @@
 //! Helpers for manipulating built-in streams
+
+pub(crate) mod stream_backoff;
+
 use crate::watcher;
 use futures::{
     pin_mut,
