+++ kube-runtime/src/watcher.rs
@@ -1,18 +1,17 @@
 //! Watches a Kubernetes Resource for changes, with error recovery
 
+use crate::utils::stream_backoff::StreamBackoff;
 use backoff::backoff::Backoff;
 use derivative::Derivative;
-use futures::{stream::BoxStream, Future, Stream, StreamExt, TryStream};
+use futures::{stream::BoxStream, Stream, StreamExt};
 use kube_client::{
     api::{ListParams, Resource, ResourceExt, WatchEvent},
     Api,
 };
-use pin_project::pin_project;
 use serde::de::DeserializeOwned;
 use smallvec::SmallVec;
-use std::{clone::Clone, fmt::Debug, pin::Pin, task::Poll};
+use std::{clone::Clone, fmt::Debug};
 use thiserror::Error;
-use tokio::time::{sleep, Instant, Sleep};
 
 #[derive(Debug, Error)]
 pub enum Error {
@@ -275,75 +274,6 @@ pub fn watch_object<K: Resource + Clone + DeserializeOwned + Debug + Send + 'sta
     //.filter(|r| ready(!std::matches!(r, Err(Error::BackoffRetriable))))
 }
 
-#[pin_project]
-struct StreamBackoff<S, B> {
-    #[pin]
-    stream: S,
-    backoff: B,
-    #[pin]
-    state: StreamBackoffState,
-}
-
-#[pin_project(project = StreamBackoffStateProj)]
-// It's expected to have relatively few but long-lived `StreamBackoff`s in a project, so we would rather have
-// cheaper sleeps than a smaller `StreamBackoff`.
-#[allow(clippy::large_enum_variant)]
-enum StreamBackoffState {
-    BackingOff(#[pin] Sleep),
-    GivenUp,
-    Awake,
-}
-
-impl<S: TryStream, B: Backoff> Stream for StreamBackoff<S, B> {
-    type Item = Result<S::Ok, S::Error>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut std::task::Context<'_>) -> Poll<Option<Self::Item>> {
-        let mut this = self.project();
-        match this.state.as_mut().project() {
-            StreamBackoffStateProj::BackingOff(mut backoff_sleep) => match backoff_sleep.as_mut().poll(cx) {
-                Poll::Ready(()) => {
-                    tracing::debug!(deadline = ?backoff_sleep.deadline(), "Backoff complete, waking up");
-                    this.state.set(StreamBackoffState::Awake)
-                }
-                Poll::Pending => {
-                    let deadline = backoff_sleep.deadline();
-                    tracing::trace!(
-                        ?deadline,
-                        remaining_duration = ?deadline.saturating_duration_since(Instant::now()),
-                        "Still waiting for backoff sleep to complete"
-                    );
-                    return Poll::Pending;
-                }
-            },
-            StreamBackoffStateProj::GivenUp => {
-                tracing::debug!("Backoff has given up, stream is closed");
-                return Poll::Ready(None);
-            }
-            StreamBackoffStateProj::Awake => {}
-        }
-
-        let next_item = this.stream.try_poll_next(cx);
-        if let Poll::Ready(Some(Err(_))) = &next_item {
-            if let Some(backoff_duration) = this.backoff.next_backoff() {
-                let backoff_sleep = sleep(backoff_duration);
-                tracing::debug!(
-                    deadline = ?backoff_sleep.deadline(),
-                    duration = ?backoff_duration,
-                    "Error received, backing off"
-                );
-                this.state.set(StreamBackoffState::BackingOff(backoff_sleep));
-            } else {
-                tracing::debug!("Error received, giving up");
-                this.state.set(StreamBackoffState::GivenUp);
-            }
-        } else {
-            tracing::trace!("Non-error received, resetting backoff");
-            this.backoff.reset();
-        }
-        next_item
-    }
-}
-
 pub fn backoff_watch<K, B>(
     stream: impl Stream<Item = Result<Event<K>>> + Send,
     backoff: B,
@@ -352,9 +282,5 @@ where
     K: Resource + Send,
     B: Backoff + Send,
 {
-    StreamBackoff {
-        stream,
-        backoff,
-        state: StreamBackoffState::Awake,
-    }
+    StreamBackoff::new(stream, backoff)
 }
